pyproject.toml
[project]
name = "rss-alter"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "being", email = "34680324+being24@users.noreply.github.com" }
]
dependencies = [
    "wikidot>=3.0.7",
    "sqlalchemy>=2.0.31",
    "pydantic>=2.8.2",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/rss_alter"]
